+++ src/ngx-drag-scroll.spec.ts
@@ -4,7 +4,7 @@ import {
   Renderer2,
   ElementRef
 } from '@angular/core';
-import { DragScroll } from './ngx-drag-scroll';
+import { DragScrollDirective } from './ngx-drag-scroll';
 
 import {
   By
@@ -23,17 +23,17 @@ import {
 class TestComponent {
 }
 
-describe('Directive: DragScroll', () => {
+describe('Directive: DragScrollDirective', () => {
   const scrollbarWidth = '15px';
   beforeEach(() => {
     TestBed.configureTestingModule({
-      declarations: [TestComponent, DragScroll]
+      declarations: [TestComponent, DragScrollDirective]
     });
   });
 
   it('should drag to scroll horizontally and vertically', async(() => {
     TestBed.overrideComponent(TestComponent, {set: {
-      template: `<div style="width: 50px; height: 50px;" drag-scroll>
+      template: `<div style="width: 50px; height: 50px;" dragScroll>
                   <div style="width: 300px; height: 300px;"></div>
                 </div>`
     }});
@@ -41,7 +41,7 @@ describe('Directive: DragScroll', () => {
     TestBed.compileComponents().then(() => {
       const fixture = TestBed.createComponent(TestComponent);
       fixture.detectChanges();
-      const compiled = fixture.debugElement.query(By.directive(DragScroll));
+      const compiled = fixture.debugElement.query(By.directive(DragScrollDirective));
 
 
       compiled.triggerEventHandler('mousedown', new MouseEvent('mousedown'));
@@ -60,7 +60,7 @@ describe('Directive: DragScroll', () => {
 
   it('should disable drag and scroll horizontally and vertically', async(() => {
     TestBed.overrideComponent(TestComponent, {set: {
-      template: `<div style="width: 50px; height: 50px;" drag-scroll drag-scroll-disabled="true">
+      template: `<div style="width: 50px; height: 50px;" dragScroll drag-scroll-disabled="true">
                   <div style="width: 300px; height: 300px;"></div>
                 </div>`
     }});
@@ -68,7 +68,7 @@ describe('Directive: DragScroll', () => {
     TestBed.compileComponents().then(() => {
       const fixture = TestBed.createComponent(TestComponent);
       fixture.detectChanges();
-      const compiled = fixture.debugElement.query(By.directive(DragScroll));
+      const compiled = fixture.debugElement.query(By.directive(DragScrollDirective));
 
       compiled.triggerEventHandler('mousedown', new MouseEvent('mousedown'));
       document.dispatchEvent(new MouseEvent('mousemove', {bubbles: true, clientX: -100}));
@@ -88,7 +88,7 @@ describe('Directive: DragScroll', () => {
 
   it('should disable drag and scroll horizontally', async(() => {
     TestBed.overrideComponent(TestComponent, {set: {
-      template: `<div style="width: 50px; height: 50px;" drag-scroll drag-scroll-x-disabled="true">
+      template: `<div style="width: 50px; height: 50px;" dragScroll drag-scroll-x-disabled="true">
                   <div style="width: 300px; height: 300px;"></div>
                 </div>`
     }});
@@ -96,7 +96,7 @@ describe('Directive: DragScroll', () => {
     TestBed.compileComponents().then(() => {
       const fixture = TestBed.createComponent(TestComponent);
       fixture.detectChanges();
-      const compiled = fixture.debugElement.query(By.directive(DragScroll));
+      const compiled = fixture.debugElement.query(By.directive(DragScrollDirective));
 
       compiled.triggerEventHandler('mousedown', new MouseEvent('mousedown'));
       document.dispatchEvent(new MouseEvent('mousemove', {bubbles: true, clientX: -100}));
@@ -116,7 +116,7 @@ describe('Directive: DragScroll', () => {
 
   it('should disable drag and scroll horizontally', async(() => {
     TestBed.overrideComponent(TestComponent, {set: {
-      template: `<div style="width: 50px; height: 50px;" drag-scroll drag-scroll-y-disabled="true">
+      template: `<div style="width: 50px; height: 50px;" dragScroll drag-scroll-y-disabled="true">
                   <div style="width: 300px; height: 300px;"></div>
                 </div>`
     }});
@@ -124,7 +124,7 @@ describe('Directive: DragScroll', () => {
     TestBed.compileComponents().then(() => {
       const fixture = TestBed.createComponent(TestComponent);
       fixture.detectChanges();
-      const compiled = fixture.debugElement.query(By.directive(DragScroll));
+      const compiled = fixture.debugElement.query(By.directive(DragScrollDirective));
 
       compiled.triggerEventHandler('mousedown', new MouseEvent('mousedown'));
       document.dispatchEvent(new MouseEvent('mousemove', {bubbles: true, clientX: -100}));
@@ -144,14 +144,14 @@ describe('Directive: DragScroll', () => {
 
   it('should only hide horizontal scroll bar', async(() => {
     TestBed.overrideComponent(TestComponent, {set: {
-      template: `<div style="width: 50px; height: 350px;" drag-scroll scrollbar-hidden="true">
+      template: `<div style="width: 50px; height: 350px;" dragScroll scrollbar-hidden="true">
                   <div style="width: 300px; height: 300px;"></div>
                 </div>`
     }});
     TestBed.compileComponents().then(() => {
       const fixture = TestBed.createComponent(TestComponent);
       fixture.detectChanges();
-      const compiled = fixture.debugElement.query(By.directive(DragScroll));
+      const compiled = fixture.debugElement.query(By.directive(DragScrollDirective));
       expect(compiled.nativeElement.style.width).toBe('100%');
       expect(compiled.nativeElement.style.height).toBe(`calc(100% + ${scrollbarWidth})`);
     });
@@ -159,14 +159,14 @@ describe('Directive: DragScroll', () => {
 
   it('should only hide vertical scroll bar', async(() => {
     TestBed.overrideComponent(TestComponent, {set: {
-      template: `<div style="width: 350px; height: 50px;" drag-scroll scrollbar-hidden="true">
+      template: `<div style="width: 350px; height: 50px;" dragScroll scrollbar-hidden="true">
                   <div style="width: 300px; height: 300px;"></div>
                 </div>`
     }});
     TestBed.compileComponents().then(() => {
       const fixture = TestBed.createComponent(TestComponent);
       fixture.detectChanges();
-      const compiled = fixture.debugElement.query(By.directive(DragScroll));
+      const compiled = fixture.debugElement.query(By.directive(DragScrollDirective));
       expect(compiled.nativeElement.style.width).toBe(`calc(100% + ${scrollbarWidth})`);
       expect(compiled.nativeElement.style.height).toBe('100%');
     });
@@ -174,7 +174,7 @@ describe('Directive: DragScroll', () => {
 
   it('should hide all scroll bars', async(() => {
     TestBed.overrideComponent(TestComponent, {set: {
-      template: `<div style="width: 50px; height: 50px;" drag-scroll scrollbar-hidden="true">
+      template: `<div style="width: 50px; height: 50px;" dragScroll scrollbar-hidden="true">
                   <div style="width: 300px; height: 300px;"></div>
                 </div>`
     }});
@@ -182,7 +182,7 @@ describe('Directive: DragScroll', () => {
       const fixture = TestBed.createComponent(TestComponent);
       fixture.detectChanges();
 
-      const compiled = fixture.debugElement.query(By.directive(DragScroll));
+      const compiled = fixture.debugElement.query(By.directive(DragScrollDirective));
       expect(compiled.nativeElement.style.width).toBe(`calc(100% + ${scrollbarWidth})`);
       expect(compiled.nativeElement.style.height).toBe(`calc(100% + ${scrollbarWidth})`);
     });
@@ -190,7 +190,7 @@ describe('Directive: DragScroll', () => {
 
   it('should not trying to hide the scrollbar when there are nothing to hide', async(() => {
     TestBed.overrideComponent(TestComponent, {set: {
-      template: `<div style="width: 50px; height: 50px;" drag-scroll>
+      template: `<div style="width: 50px; height: 50px;" dragScroll>
                   <div style="width: 49px; height: 49px;"></div>
                 </div>`
     }});
@@ -198,7 +198,7 @@ describe('Directive: DragScroll', () => {
       const fixture = TestBed.createComponent(TestComponent);
       fixture.detectChanges();
 
-      const compiled = fixture.debugElement.query(By.directive(DragScroll));
+      const compiled = fixture.debugElement.query(By.directive(DragScrollDirective));
       expect(compiled.nativeElement.style.width).toBe('50px');
       expect(compiled.nativeElement.style.height).toBe('50px');
     });